Fix RabbitMQ documentation
The RabbitMQ documentation was lacking when RabbitMQ was used but messages arrived in the "Unacked" state rather than the "Ready" state, which can happens in frameworks such as Dapr.

This commit clarifies that this can be resolved through the HTTP interface and also points to the correct endpoint to utilize.

Signed-off-by: Xavier Geerinck <xavier.geerinck@gmail.com>