Added initialDelayCooldownPeriod parameter to KEDA ScaledObject
This commit introduces the initialDelayCooldownPeriod parameter for KEDA's ScaledObject resource. This parameter sets a delay before the cooldownPeriod activates after the initial creation of the ScaledObject. By default, this delay is 0 seconds. If configured, e.g., to 120 seconds, the cooldown period only begins after the ScaledObject has been active for the specified duration. This feature allows for more controlled scaling behavior during the initial deployment phase.

Signed-off-by: wangxingrui <xrwang8@gamil.com>